feat: introduce backend abstraction layer
Abstract calls away from the GH cli through an indirection.
Pick backend cli program based upon remote host link.

This seems like the most sensible approach since it maintains type hints
through luals.
