+++ lua/octo/backend/gh/init.lua
@@ -1,0 +1,102 @@
+local cli = require "octo.backend.gh.cli"
+local commands = require "octo.backend.gh.commands"
+local telescope_commands = require "octo.backend.gh.telescope-commands"
+local fzf_lua_commands = require "octo.backend.gh.fzf-lua-commands"
+
+local M = {}
+
+M.functions = {
+  ["setup"] = cli.setup,
+  ["load"] = commands.load,
+  ["reactions_popup"] = commands.reactions_popup,
+  ["user_popup"] = commands.user_popup,
+  ["link_popup"] = commands.link_popup,
+  ["cmds_delete_comment"] = commands.cmds_delete_comment,
+  ["cmds_resolve_thread"] = commands.cmds_resolve_thread,
+  ["cmds_unresolve_thread"] = commands.cmds_unresolve_thread,
+  ["cmds_change_state"] = commands.cmds_change_state,
+  ["cmds_save_issue"] = commands.cmds_save_issue,
+  ["cmds_save_pr"] = commands.cmds_save_pr,
+  ["cmds_mark_pr_ready"] = commands.cmds_mark_pr_ready,
+  ["cmds_mark_pr_draft"] = commands.cmds_mark_pr_draft,
+  ["cmds_pr_checks"] = commands.cmds_pr_checks,
+  ["cmds_merge_pr"] = commands.cmds_merge_pr,
+  ["cmds_show_pr_diff"] = commands.cmds_show_pr_diff,
+  ["cmds_reaction_action"] = commands.cmds_reaction_action,
+  ["cmds_add_project_card"] = commands.cmds_add_project_card,
+  ["cmds_remove_project_card"] = commands.cmds_remove_project_card,
+  ["cmds_move_project_card"] = commands.cmds_move_project_card,
+  ["cmds_set_project_card_v2"] = commands.cmds_set_project_card_v2,
+  ["cmds_remove_project_card_v2"] = commands.cmds_remove_project_card_v2,
+  ["cmds_create_label"] = commands.cmds_create_label,
+  ["cmds_add_label"] = commands.cmds_add_label,
+  ["cmds_remove_label"] = commands.cmds_remove_label,
+  ["cmds_add_user"] = commands.cmds_add_user,
+  ["cmds_remove_assignee"] = commands.cmds_remove_assignee,
+  ["buffer_fetch_taggable_users"] = commands.buffer_fetch_taggable_users,
+  ["buffer_fetch_issues"] = commands.buffer_fetch_issues,
+  ["buffer_save_title_and_body"] = commands.buffer_save_title_and_body,
+  ["buffer_add_issue_comment"] = commands.buffer_add_issue_comment,
+  ["buffer_add_thread_comment"] = commands.buffer_add_thread_comment,
+  ["buffer_pr_add_thread"] = commands.buffer_pr_add_thread,
+  ["buffer_commit_add_thread"] = commands.buffer_commit_add_thread,
+  ["buffer_add_pr_comment"] = commands.buffer_add_pr_comment,
+  ["buffer_update_comment"] = commands.buffer_update_comment,
+  ["pr_get_diff"] = commands.pr_get_diff,
+  ["pr_get_changed_files"] = commands.pr_get_changed_files,
+  ["pr_get_commit_changed_files"] = commands.pr_get_commit_changed_files,
+  ["go_to_issue"] = commands.go_to_issue,
+  ["file_toggle_viewed"] = commands.file_toggle_viewed,
+  ["review_start_review_mutation"] = commands.review_start_review_mutation,
+  ["review_retrieve"] = commands.review_retrieve,
+  ["review_discard"] = commands.review_discard,
+  ["review_submit"] = commands.review_submit,
+  ["util_checkout_pr"] = commands.util_checkout_pr,
+  ["util_checkout_pr_sync"] = commands.util_checkout_pr_sync,
+  ["util_merge_pr"] = commands.util_merge_pr,
+  ["util_get_repo_iid"] = commands.util_get_repo_iid,
+  ["util_get_repo_info"] = commands.util_get_repo_info,
+  ["util_get_repo_templates"] = commands.util_get_repo_templates,
+  ["util_get_file_contents"] = commands.util_get_file_contents,
+  ["util_fork"] = commands.util_fork,
+  ["util_get_pr_for_curr_branch"] = commands.util_get_pr_for_curr_branch,
+  ["util_get_user_id"] = commands.util_get_user_id,
+  ["util_get_label_id"] = commands.util_get_label_id,
+  ["get_user_name"] = commands.get_user_name,
+  ["open_in_browser"] = commands.open_in_browser,
+  ["telescope_default_issue"] = telescope_commands.telescope_default_issue,
+  ["telescope_default_commit"] = telescope_commands.telescope_default_commit,
+  ["telescope_issues"] = telescope_commands.telescope_issues,
+  ["telescope_gists"] = telescope_commands.telescope_gists,
+  ["telescope_pull_requests"] = telescope_commands.telescope_pull_requests,
+  ["telescope_commits"] = telescope_commands.telescope_commits,
+  ["telescope_review_commits"] = telescope_commands.telescope_review_commits,
+  ["telescope_changed_files"] = telescope_commands.telescope_changed_files,
+  ["telescope_search"] = telescope_commands.telescope_search,
+  ["telescope_select_target_project_column"] = telescope_commands.telescope_select_target_project_column,
+  ["telescope_select_label"] = telescope_commands.telescope_select_label,
+  ["telescope_select_assigned_label"] = telescope_commands.telescope_select_assigned_label,
+  ["telescope_get_users"] = telescope_commands.telescope_get_users,
+  ["telescope_select_assignee"] = telescope_commands.telescope_select_assignee,
+  ["telescope_repos"] = telescope_commands.telescope_repos,
+  ["fzf_lua_default_issue"] = fzf_lua_commands.fzf_lua_default_issue,
+  ["fzf_lua_previewer_search"] = fzf_lua_commands.fzf_lua_previewer_search,
+  ["fzf_lua_default_commit"] = fzf_lua_commands.fzf_lua_default_commit,
+  ["fzf_lua_previewer_repos"] = fzf_lua_commands.fzf_lua_previewer_repos,
+  ["fzf_lua_issues"] = fzf_lua_commands.fzf_lua_issues,
+  ["fzf_lua_gists"] = fzf_lua_commands.fzf_lua_gists,
+  ["fzf_lua_pull_requests"] = fzf_lua_commands.fzf_lua_pull_requests,
+  ["fzf_lua_commits"] = fzf_lua_commands.fzf_lua_commits,
+  ["fzf_lua_review_commits"] = fzf_lua_commands.fzf_lua_review_commits,
+  ["fzf_lua_changed_files"] = fzf_lua_commands.fzf_lua_changed_files,
+  ["fzf_lua_search"] = fzf_lua_commands.fzf_lua_search,
+  ["fzf_lua_select_target_project_column"] = fzf_lua_commands.fzf_lua_select_target_project_column,
+  ["fzf_lua_select_target_project_column_v2"] = fzf_lua_commands.fzf_lua_select_target_project_column_v2,
+  ["fzf_lua_select_label"] = fzf_lua_commands.fzf_lua_select_label,
+  ["fzf_lua_select_assigned_label"] = fzf_lua_commands.fzf_lua_select_assigned_label,
+  ["fzf_lua_get_users"] = fzf_lua_commands.fzf_lua_get_users,
+  ["fzf_lua_select_assignee"] = fzf_lua_commands.fzf_lua_select_assignee,
+  ["fzf_lua_repos"] = fzf_lua_commands.fzf_lua_repos,
+}
+
+return M
