+++ lua/octo/init.lua
@@ -5,13 +5,10 @@ local constants = require "octo.constants"
 local commands = require "octo.commands"
 local completion = require "octo.completion"
 local folds = require "octo.folds"
-local gh = require "octo.gh"
-local graphql = require "octo.gh.graphql"
+local backend = require "octo.backend"
 local picker = require "octo.picker"
 local reviews = require "octo.reviews"
 local signs = require "octo.ui.signs"
-local window = require "octo.ui.window"
-local writers = require "octo.ui.writers"
 local utils = require "octo.utils"
 local vim = vim
 
@@ -28,8 +25,7 @@ function M.setup(user_config)
   end
 
   config.setup(user_config or {})
-  if not vim.fn.executable(config.values.gh_cmd) then
-    utils.error("gh executable not found using path: " .. config.values.gh_cmd)
+  if not backend.available_executable() then
     return
   end
 
@@ -39,7 +35,8 @@ function M.setup(user_config)
   folds.setup()
   autocmds.setup()
   commands.setup()
-  gh.setup()
+  local func = backend.get_funcs()["setup"]
+  func()
 end
 
 function M.configure_octo_buffer(bufnr)
@@ -89,36 +86,8 @@ function M.load_buffer(bufnr)
 end
 
 function M.load(repo, kind, number, cb)
-  local owner, name = utils.split_repo(repo)
-  local query, key
-
-  if kind == "pull" then
-    query = graphql("pull_request_query", owner, name, number, _G.octo_pv2_fragment)
-    key = "pullRequest"
-  elseif kind == "issue" then
-    query = graphql("issue_query", owner, name, number, _G.octo_pv2_fragment)
-    key = "issue"
-  elseif kind == "repo" then
-    query = graphql("repository_query", owner, name)
-  end
-  gh.run {
-    args = { "api", "graphql", "--paginate", "--jq", ".", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        vim.api.nvim_err_writeln(stderr)
-      elseif output then
-        if kind == "pull" or kind == "issue" then
-          local resp = utils.aggregate_pages(output, string.format("data.repository.%s.timelineItems.nodes", key))
-          local obj = resp.data.repository[key]
-          cb(obj)
-        elseif kind == "repo" then
-          local resp = vim.fn.json_decode(output)
-          local obj = resp.data.repository
-          cb(obj)
-        end
-      end
-    end,
-  }
+  local func = backend.get_funcs()["load"]
+  func(repo, kind, number, cb)
 end
 
 function M.render_signs()
@@ -137,61 +106,16 @@ function M.on_cursor_hold()
   -- reactions popup
   local id = buffer:get_reactions_at_cursor()
   if id then
-    local query = graphql("reactions_for_object_query", id)
-    gh.run {
-      args = { "api", "graphql", "-f", string.format("query=%s", query) },
-      cb = function(output, stderr)
-        if stderr and not utils.is_blank(stderr) then
-          vim.api.nvim_err_writeln(stderr)
-        elseif output then
-          local resp = vim.fn.json_decode(output)
-          local reactions = {}
-          local reactionGroups = resp.data.node.reactionGroups
-          for _, reactionGroup in ipairs(reactionGroups) do
-            local users = reactionGroup.users.nodes
-            local logins = {}
-            for _, user in ipairs(users) do
-              table.insert(logins, user.login)
-            end
-            if #logins > 0 then
-              reactions[reactionGroup.content] = logins
-            end
-          end
-          local popup_bufnr = vim.api.nvim_create_buf(false, true)
-          local lines_count, max_length = writers.write_reactions_summary(popup_bufnr, reactions)
-          window.create_popup {
-            bufnr = popup_bufnr,
-            width = 4 + max_length,
-            height = 2 + lines_count,
-          }
-        end
-      end,
-    }
+    local func = backend.get_funcs()["reactions_popup"]
+    func(id)
     return
   end
 
   -- user popup
   local login = utils.extract_pattern_at_cursor(constants.USER_PATTERN)
   if login then
-    local query = graphql("user_profile_query", login)
-    gh.run {
-      args = { "api", "graphql", "-f", string.format("query=%s", query) },
-      cb = function(output, stderr)
-        if stderr and not utils.is_blank(stderr) then
-          vim.api.nvim_err_writeln(stderr)
-        elseif output then
-          local resp = vim.fn.json_decode(output)
-          local user = resp.data.user
-          local popup_bufnr = vim.api.nvim_create_buf(false, true)
-          local lines, max_length = writers.write_user_profile(popup_bufnr, user)
-          window.create_popup {
-            bufnr = popup_bufnr,
-            width = 4 + max_length,
-            height = 2 + lines,
-          }
-        end
-      end,
-    }
+    local func = backend.get_funcs()["user_popup"]
+    func(login)
     return
   end
 
@@ -200,27 +124,8 @@ function M.on_cursor_hold()
   if not repo or not number then
     return
   end
-  local owner, name = utils.split_repo(repo)
-  local query = graphql("issue_summary_query", owner, name, number)
-  gh.run {
-    args = { "api", "graphql", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        vim.api.nvim_err_writeln(stderr)
-      elseif output then
-        local resp = vim.fn.json_decode(output)
-        local issue = resp.data.repository.issueOrPullRequest
-        local popup_bufnr = vim.api.nvim_create_buf(false, true)
-        local max_length = 80
-        local lines = writers.write_issue_summary(popup_bufnr, issue, { max_length = max_length })
-        window.create_popup {
-          bufnr = popup_bufnr,
-          width = max_length,
-          height = 2 + lines,
-        }
-      end
-    end,
-  }
+  local func = backend.get_funcs()["link_popup"]
+  func(repo, number)
 end
 
 function M.create_buffer(kind, obj, repo, create)
