+++ lua/octo/model/pull-request.lua
@@ -1,5 +1,5 @@
 local utils = require "octo.utils"
-local gh = require "octo.gh"
+local backend = require "octo.backend"
 
 local M = {}
 
@@ -59,86 +59,20 @@ M.PullRequest = PullRequest
 
 ---Fetch the diff of the PR
 function PullRequest:get_diff(pr)
-  local url = string.format("repos/%s/pulls/%d", pr.repo, pr.number)
-  gh.run {
-    args = { "api", "--paginate", url },
-    headers = { "Accept: application/vnd.github.v3.diff" },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        pr.diff = output
-      end
-    end,
-  }
+  local func = backend.get_funcs()["pr_get_diff"]
+  pr.diff = func(pr.repo, pr.number)
 end
 
 ---Fetch the changed files for a given PR
 function PullRequest:get_changed_files(callback)
-  local url = string.format("repos/%s/pulls/%d/files", self.repo, self.number)
-  gh.run {
-    args = { "api", "--paginate", url, "--jq", "." },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local FileEntry = require("octo.reviews.file-entry").FileEntry
-        local results = vim.fn.json_decode(output)
-        local files = {}
-        for _, result in ipairs(results) do
-          local entry = FileEntry:new {
-            path = result.filename,
-            previous_path = result.previous_filename,
-            patch = result.patch,
-            pull_request = self,
-            status = utils.file_status_map[result.status],
-            stats = {
-              additions = result.additions,
-              deletions = result.deletions,
-              changes = result.changes,
-            },
-          }
-          table.insert(files, entry)
-        end
-        callback(files)
-      end
-    end,
-  }
+  local func = backend.get_funcs()["pr_get_changed_files"]
+  func(self, callback)
 end
 
 ---Fetch the changed files at a given commit
 function PullRequest:get_commit_changed_files(rev, callback)
-  local url = string.format("repos/%s/commits/%s", self.repo, rev.commit)
-  gh.run {
-    args = { "api", "--paginate", url, "--jq", "." },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local FileEntry = require("octo.reviews.file-entry").FileEntry
-        local results = vim.fn.json_decode(output)
-        local files = {}
-        if results.files then
-          for _, result in ipairs(results.files) do
-            local entry = FileEntry:new {
-              path = result.filename,
-              previous_path = result.previous_filename,
-              patch = result.patch,
-              pull_request = self,
-              status = utils.file_status_map[result.status],
-              stats = {
-                additions = result.additions,
-                deletions = result.deletions,
-                changes = result.changes,
-              },
-            }
-            table.insert(files, entry)
-          end
-          callback(files)
-        end
-      end
-    end,
-  }
+  local func = backend.get_funcs()["pr_get_commit_changed_files"]
+  func(self, rev, callback)
 end
 
 return M
