+++ lua/octo/pickers/fzf-lua/pickers/issues.lua
@@ -1,8 +1,5 @@
 local fzf_actions = require "octo.pickers.fzf-lua.pickers.fzf_actions"
-local entry_maker = require "octo.pickers.fzf-lua.entry_maker"
 local fzf = require "fzf-lua"
-local gh = require "octo.gh"
-local graphql = require "octo.gh.graphql"
 local octo_config = require "octo.config"
 local picker_utils = require "octo.pickers.fzf-lua.pickers.utils"
 local previewers = require "octo.pickers.fzf-lua.previewers"
@@ -22,48 +19,14 @@ return function(opts)
     return
   end
 
-  local owner, name = utils.split_repo(opts.repo)
   local cfg = octo_config.values
   local order_by = cfg.issues.order_by
-
-  local query = graphql("issues_query", owner, name, filter, order_by.field, order_by.direction, { escape = false })
-
   local formatted_issues = {}
 
   local get_contents = function(fzf_cb)
-    gh.run {
-      args = {
-        "api",
-        "graphql",
-        "--paginate",
-        "--jq",
-        ".",
-        "-f",
-        string.format("query=%s", query),
-      },
-      stream_cb = function(data, err)
-        if err and not utils.is_blank(err) then
-          utils.error(err)
-          fzf_cb()
-        elseif data then
-          local resp = utils.aggregate_pages(data, "data.repository.issues.nodes")
-          local issues = resp.data.repository.issues.nodes
-
-          for _, issue in ipairs(issues) do
-            local entry = entry_maker.gen_from_issue(issue)
-
-            if entry ~= nil then
-              formatted_issues[entry.ordinal] = entry
-              local prefix = fzf.utils.ansi_from_hl("Comment", entry.value)
-              fzf_cb(prefix .. " " .. entry.obj.title)
-            end
-          end
-        end
-      end,
-      cb = function()
-        fzf_cb()
-      end,
-    }
+    local backend = require "octo.backend"
+    local func = backend.get_funcs()["fzf_lua_issues"]
+    func(formatted_issues, opts.repo, order_by, filter, fzf_cb)
   end
 
   fzf.fzf_exec(get_contents, {
