+++ lua/octo/pickers/fzf-lua/pickers/labels.lua
@@ -1,8 +1,5 @@
 local fzf = require "fzf-lua"
-local gh = require "octo.gh"
-local graphql = require "octo.gh.graphql"
 local picker_utils = require "octo.pickers.fzf-lua.pickers.utils"
-local utils = require "octo.utils"
 
 return function(cb)
   local bufnr = vim.api.nvim_get_current_buf()
@@ -12,27 +9,10 @@ return function(cb)
     return
   end
 
-  local query = graphql("labels_query", buffer.owner, buffer.name)
-
   local get_contents = function(fzf_cb)
-    gh.run {
-      args = { "api", "graphql", "-f", string.format("query=%s", query) },
-      cb = function(output, stderr)
-        if stderr and not utils.is_blank(stderr) then
-          utils.error(stderr)
-        elseif output then
-          local resp = vim.fn.json_decode(output)
-          local labels = resp.data.repository.labels.nodes
-
-          for _, label in ipairs(labels) do
-            local colored_name = picker_utils.color_string_with_hex(label.name, "#" .. label.color)
-            fzf_cb(string.format("%s %s", label.id, colored_name))
-          end
-        end
-
-        fzf_cb()
-      end,
-    }
+    local backend = require "octo.backend"
+    local func = backend.get_funcs()["fzf_lua_select_label"]
+    func(buffer, fzf_cb)
   end
 
   fzf.fzf_exec(
