+++ lua/octo/pickers/fzf-lua/pickers/project_columns.lua
@@ -1,9 +1,6 @@
 local entry_maker = require "octo.pickers.fzf-lua.entry_maker"
 local fzf = require "fzf-lua"
-local gh = require "octo.gh"
-local graphql = require "octo.gh.graphql"
 local picker_utils = require "octo.pickers.fzf-lua.pickers.utils"
-local utils = require "octo.utils"
 
 return function(cb)
   local bufnr = vim.api.nvim_get_current_buf()
@@ -21,39 +18,9 @@ return function(cb)
   })
 
   local get_projects = function(fzf_cb)
-    local query = graphql("projects_query", buffer.owner, buffer.name, vim.g.octo_viewer, buffer.owner)
-    gh.run {
-      args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
-      cb = function(output)
-        if output then
-          local resp = vim.fn.json_decode(output)
-
-          local projects = {}
-          local user_projects = resp.data.user and resp.data.user.projects.nodes or {}
-          local repo_projects = resp.data.repository and resp.data.repository.projects.nodes or {}
-          local org_projects = not resp.errors and resp.data.organization.projects.nodes or {}
-          vim.list_extend(projects, repo_projects)
-          vim.list_extend(projects, user_projects)
-          vim.list_extend(projects, org_projects)
-
-          if #projects == 0 then
-            utils.error(string.format("There are no matching projects for %s.", buffer.repo))
-            fzf_cb()
-          end
-
-          for _, project in ipairs(projects) do
-            local entry = entry_maker.gen_from_project(project)
-
-            if entry ~= nil then
-              formatted_projects[entry.ordinal] = entry
-              fzf_cb(entry.ordinal)
-            end
-          end
-        end
-
-        fzf_cb()
-      end,
-    }
+    local backend = require "octo.backend"
+    local func = backend.get_funcs()["fzf_lua_select_target_project_column"]
+    func(formatted_projects, buffer, fzf_cb)
   end
 
   local default_action = function(selected_project)
