+++ lua/octo/pickers/telescope/provider.lua
@@ -1,5 +1,4 @@
-local gh = require "octo.gh"
-local graphql = require "octo.gh.graphql"
+local backend = require "octo.backend"
 local navigation = require "octo.navigation"
 local previewers = require "octo.pickers.telescope.previewers"
 local entry_maker = require "octo.pickers.telescope.entry_maker"
@@ -135,53 +134,12 @@ function M.issues(opts)
     return
   end
 
-  local owner, name = utils.split_repo(opts.repo)
-  local cfg = octo_config.values
-  local order_by = cfg.issues.order_by
-  local query = graphql("issues_query", owner, name, filter, order_by.field, order_by.direction, { escape = false })
-  utils.info "Fetching issues (this may take a while) ..."
-  gh.run {
-    args = { "api", "graphql", "--paginate", "--jq", ".", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local resp = utils.aggregate_pages(output, "data.repository.issues.nodes")
-        local issues = resp.data.repository.issues.nodes
-        if #issues == 0 then
-          utils.error(string.format("There are no matching issues in %s.", opts.repo))
-          return
-        end
-        local max_number = -1
-        for _, issue in ipairs(issues) do
-          if #tostring(issue.number) > max_number then
-            max_number = #tostring(issue.number)
-          end
-        end
-        opts.preview_title = opts.preview_title or ""
-        opts.prompt_title = opts.prompt_title or ""
-        opts.results_title = opts.results_title or ""
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = issues,
-              entry_maker = entry_maker.gen_from_issue(max_number),
-            },
-            sorter = conf.generic_sorter(opts),
-            previewer = previewers.issue.new(opts),
-            attach_mappings = function(_, map)
-              action_set.select:replace(function(prompt_bufnr, type)
-                open(type)(prompt_bufnr)
-              end)
-              map("i", cfg.picker_config.mappings.open_in_browser.lhs, open_in_browser())
-              map("i", cfg.picker_config.mappings.copy_url.lhs, copy_url())
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  opts.open = open
+  opts.open_in_browser = open_in_browser
+  opts.copy_url = copy_url
+
+  local func = backend.get_funcs()["telescope_issues"]
+  func(opts, octo_config.values, filter)
 end
 
 --
@@ -215,35 +173,11 @@ function M.gists(opts)
   else
     privacy = "ALL"
   end
-  local query = graphql("gists_query", privacy)
-  gh.run {
-    args = { "api", "graphql", "--paginate", "--jq", ".", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local resp = utils.aggregate_pages(output, "data.viewer.gists.nodes")
-        local gists = resp.data.viewer.gists.nodes
-        opts.preview_title = opts.preview_title or ""
-        opts.prompt_title = opts.prompt_title or ""
-        opts.results_title = opts.results_title or ""
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = gists,
-              entry_maker = entry_maker.gen_from_gist(),
-            },
-            previewer = previewers.gist.new(opts),
-            sorter = conf.generic_sorter(opts),
-            attach_mappings = function(_, map)
-              map("i", "<CR>", open_gist)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+
+  opts.open_gist = open_gist
+
+  local func = backend.get_funcs()["telescope_gists"]
+  func(opts, privacy)
 end
 
 --
@@ -280,56 +214,14 @@ function M.pull_requests(opts)
     return
   end
 
-  local owner, name = utils.split_repo(opts.repo)
-  local cfg = octo_config.values
-  local order_by = cfg.pull_requests.order_by
-  local query =
-    graphql("pull_requests_query", owner, name, filter, order_by.field, order_by.direction, { escape = false })
-  utils.info "Fetching pull requests (this may take a while) ..."
-  gh.run {
-    args = { "api", "graphql", "--paginate", "--jq", ".", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local resp = utils.aggregate_pages(output, "data.repository.pullRequests.nodes")
-        local pull_requests = resp.data.repository.pullRequests.nodes
-        if #pull_requests == 0 then
-          utils.error(string.format("There are no matching pull requests in %s.", opts.repo))
-          return
-        end
-        local max_number = -1
-        for _, pull in ipairs(pull_requests) do
-          if #tostring(pull.number) > max_number then
-            max_number = #tostring(pull.number)
-          end
-        end
-        opts.preview_title = opts.preview_title or ""
-        opts.prompt_title = opts.prompt_title or ""
-        opts.results_title = opts.results_title or ""
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = pull_requests,
-              entry_maker = entry_maker.gen_from_issue(max_number),
-            },
-            sorter = conf.generic_sorter(opts),
-            previewer = previewers.issue.new(opts),
-            attach_mappings = function(_, map)
-              action_set.select:replace(function(prompt_bufnr, type)
-                open(type)(prompt_bufnr)
-              end)
-              map("i", cfg.picker_config.mappings.checkout_pr.lhs, checkout_pull_request())
-              map("i", cfg.picker_config.mappings.open_in_browser.lhs, open_in_browser())
-              map("i", cfg.picker_config.mappings.copy_url.lhs, copy_url())
-              map("i", cfg.picker_config.mappings.merge_pr.lhs, merge_pull_request())
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  opts.open = open
+  opts.checkout_pull_request = checkout_pull_request
+  opts.open_in_browser = open_in_browser
+  opts.copy_url = copy_url
+  opts.merge_pull_request = merge_pull_request
+
+  local func = backend.get_funcs()["telescope_pull_requests"]
+  func(opts, octo_config.values, filter)
 end
 
 --
@@ -341,100 +233,18 @@ function M.commits()
   if not buffer or not buffer:isPullRequest() then
     return
   end
-  -- TODO: graphql
-  local url = string.format("repos/%s/pulls/%d/commits", buffer.repo, buffer.number)
-  gh.run {
-    args = { "api", "--paginate", url },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local results = vim.fn.json_decode(output)
-        pickers
-          .new({}, {
-            prompt_title = false,
-            results_title = false,
-            preview_title = false,
-            finder = finders.new_table {
-              results = results,
-              entry_maker = entry_maker.gen_from_git_commits(),
-            },
-            sorter = conf.generic_sorter {},
-            previewer = previewers.commit.new { repo = buffer.repo },
-            attach_mappings = function()
-              action_set.select:replace(function(prompt_bufnr, type)
-                open_preview_buffer(type)(prompt_bufnr)
-              end)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  local func = backend.get_funcs()["telescope_commits"]
+  func(buffer, open_preview_buffer)
 end
 
-function M.review_commits(callback)
+function M.review_commits(cb)
   local current_review = require("octo.reviews").get_current_review()
   if not current_review then
     utils.error "No review in progress"
     return
   end
-  -- TODO: graphql
-  local url =
-    string.format("repos/%s/pulls/%d/commits", current_review.pull_request.repo, current_review.pull_request.number)
-  gh.run {
-    args = { "api", "--paginate", url },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local results = vim.fn.json_decode(output)
-
-        -- add a fake entry to represent the entire pull request
-        table.insert(results, {
-          sha = current_review.pull_request.right.commit,
-          commit = {
-            message = "[[ENTIRE PULL REQUEST]]",
-            author = {
-              name = "",
-              email = "",
-              date = "",
-            },
-          },
-          parents = {
-            {
-              sha = current_review.pull_request.left.commit,
-            },
-          },
-        })
-
-        pickers
-          .new({}, {
-            prompt_title = false,
-            results_title = false,
-            preview_title = false,
-            finder = finders.new_table {
-              results = results,
-              entry_maker = entry_maker.gen_from_git_commits(),
-            },
-            sorter = conf.generic_sorter {},
-            previewer = previewers.commit.new { repo = current_review.pull_request.repo },
-            attach_mappings = function()
-              action_set.select:replace(function(prompt_bufnr)
-                local commit = action_state.get_selected_entry(prompt_bufnr)
-                local right = commit.value
-                local left = commit.parent
-                actions.close(prompt_bufnr)
-                callback(right, left)
-              end)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  local func = backend.get_funcs()["telescope_review_commits"]
+  func(current_review, cb)
 end
 
 --
@@ -446,36 +256,8 @@ function M.changed_files()
   if not buffer or not buffer:isPullRequest() then
     return
   end
-  local url = string.format("repos/%s/pulls/%d/files", buffer.repo, buffer.number)
-  gh.run {
-    args = { "api", "--paginate", url },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local results = vim.fn.json_decode(output)
-        pickers
-          .new({}, {
-            prompt_title = false,
-            results_title = false,
-            preview_title = false,
-            finder = finders.new_table {
-              results = results,
-              entry_maker = entry_maker.gen_from_git_changed_files(),
-            },
-            sorter = conf.generic_sorter {},
-            previewer = previewers.changed_files.new { repo = buffer.repo, number = buffer.number },
-            attach_mappings = function()
-              action_set.select:replace(function(prompt_bufnr, type)
-                open_preview_buffer(type)(prompt_bufnr)
-              end)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  local func = backend.get_funcs()["telescope_changed_files"]
+  func(buffer, open_preview_buffer)
 end
 
 ---
@@ -498,11 +280,8 @@ function M.search(opts)
         if val then
           _prompt = string.format("%s %s", val, _prompt)
         end
-        local query = graphql("search_query", _prompt)
-        local output = gh.run {
-          args = { "api", "graphql", "-f", string.format("query=%s", query) },
-          mode = "sync",
-        }
+        local func = backend.get_funcs()["telescope_search"]
+        local output = func(_prompt)
         if output then
           local resp = vim.fn.json_decode(output)
           for _, issue in ipairs(resp.data.search.nodes) do
@@ -618,150 +397,31 @@ function M.select_target_project_column(cb)
   if not buffer then
     return
   end
-
-  local query = graphql("projects_query", buffer.owner, buffer.name, vim.g.octo_viewer, buffer.owner)
-  gh.run {
-    args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
-    cb = function(output)
-      if output then
-        local resp = vim.fn.json_decode(output)
-        local projects = {}
-        local user_projects = resp.data.user and resp.data.user.projects.nodes or {}
-        local repo_projects = resp.data.repository and resp.data.repository.projects.nodes or {}
-        local org_projects = not resp.errors and resp.data.organization.projects.nodes or {}
-        vim.list_extend(projects, repo_projects)
-        vim.list_extend(projects, user_projects)
-        vim.list_extend(projects, org_projects)
-        if #projects == 0 then
-          utils.error(string.format("There are no matching projects for %s.", buffer.repo))
-          return
-        end
-
-        local opts = vim.deepcopy(dropdown_opts)
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = projects,
-              entry_maker = entry_maker.gen_from_project(),
-            },
-            sorter = conf.generic_sorter(opts),
-            attach_mappings = function()
-              action_set.select:replace(function(prompt_bufnr)
-                local selected_project = action_state.get_selected_entry(prompt_bufnr)
-                actions._close(prompt_bufnr, true)
-                local opts2 = vim.deepcopy(dropdown_opts)
-                pickers
-                  .new(opts2, {
-                    finder = finders.new_table {
-                      results = selected_project.project.columns.nodes,
-                      entry_maker = entry_maker.gen_from_project_column(),
-                    },
-                    sorter = conf.generic_sorter(opts2),
-                    attach_mappings = function()
-                      action_set.select:replace(function(prompt_bufnr2)
-                        local selected_column = action_state.get_selected_entry(prompt_bufnr2)
-                        actions.close(prompt_bufnr2)
-                        cb(selected_column.column.id)
-                      end)
-                      return true
-                    end,
-                  })
-                  :find()
-              end)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  local func = backend.get_funcs()["telescope_select_target_project_column"]
+  func(buffer, dropdown_opts, cb)
 end
 
 --
 -- LABELS
 --
 function M.select_label(cb)
-  local opts = vim.deepcopy(dropdown_opts)
   local bufnr = vim.api.nvim_get_current_buf()
   local buffer = octo_buffers[bufnr]
   if not buffer then
     return
   end
-
-  local query = graphql("labels_query", buffer.owner, buffer.name)
-  gh.run {
-    args = { "api", "graphql", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local resp = vim.fn.json_decode(output)
-        local labels = resp.data.repository.labels.nodes
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = labels,
-              entry_maker = entry_maker.gen_from_label(),
-            },
-            sorter = conf.generic_sorter(opts),
-            attach_mappings = function(_, _)
-              actions.select_default:replace(function(prompt_bufnr)
-                local selected_label = action_state.get_selected_entry(prompt_bufnr)
-                actions.close(prompt_bufnr)
-                cb(selected_label.label.id)
-              end)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  local func = backend.get_funcs()["telescope_select_label"]
+  func(buffer, dropdown_opts, cb)
 end
 
 function M.select_assigned_label(cb)
-  local opts = vim.deepcopy(dropdown_opts)
   local bufnr = vim.api.nvim_get_current_buf()
   local buffer = octo_buffers[bufnr]
   if not buffer then
     return
   end
-  local query, key
-  if buffer:isIssue() then
-    query = graphql("issue_labels_query", buffer.owner, buffer.name, buffer.number)
-    key = "issue"
-  elseif buffer:isPullRequest() then
-    query = graphql("pull_request_labels_query", buffer.owner, buffer.name, buffer.number)
-    key = "pullRequest"
-  end
-  gh.run {
-    args = { "api", "graphql", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local resp = vim.fn.json_decode(output)
-        local labels = resp.data.repository[key].labels.nodes
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = labels,
-              entry_maker = entry_maker.gen_from_label(),
-            },
-            sorter = conf.generic_sorter(opts),
-            attach_mappings = function(_, _)
-              actions.select_default:replace(function(prompt_bufnr)
-                local selected_label = action_state.get_selected_entry(prompt_bufnr)
-                actions.close(prompt_bufnr)
-                cb(selected_label.label.id)
-              end)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  local func = backend.get_funcs()["telescope_select_assigned_label"]
+  func(buffer, dropdown_opts, cb)
 end
 
 --
@@ -781,11 +441,9 @@ function M.select_user(cb)
       if not prompt or prompt == "" or utils.is_blank(prompt) then
         return {}
       end
-      local query = graphql("users_query", prompt)
-      local output = gh.run {
-        args = { "api", "graphql", "--paginate", "-f", string.format("query=%s", query) },
-        mode = "sync",
-      }
+
+      local func = backend.get_funcs()["telescope_get_users"]
+      local output = func(prompt)
       if output then
         local users = {}
         local orgs = {}
@@ -879,48 +537,14 @@ end
 -- ASSIGNEES
 --
 function M.select_assignee(cb)
-  local opts = vim.deepcopy(dropdown_opts)
   local bufnr = vim.api.nvim_get_current_buf()
   local buffer = octo_buffers[bufnr]
   if not buffer then
     return
   end
-  local query, key
-  if buffer:isIssue() then
-    query = graphql("issue_assignees_query", buffer.owner, buffer.name, buffer.number)
-    key = "issue"
-  elseif buffer:isPullRequest() then
-    query = graphql("pull_request_assignees_query", buffer.owner, buffer.name, buffer.number)
-    key = "pullRequest"
-  end
-  gh.run {
-    args = { "api", "graphql", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local resp = vim.fn.json_decode(output)
-        local assignees = resp.data.repository[key].assignees.nodes
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = assignees,
-              entry_maker = entry_maker.gen_from_user(),
-            },
-            sorter = conf.generic_sorter(opts),
-            attach_mappings = function(_, _)
-              actions.select_default:replace(function(prompt_bufnr)
-                local selected_assignee = action_state.get_selected_entry(prompt_bufnr)
-                actions.close(prompt_bufnr)
-                cb(selected_assignee.user.id)
-              end)
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+
+  local func = backend.get_funcs()["telescope_select_assignee"]
+  func(buffer, dropdown_opts, cb)
 end
 
 --
@@ -928,60 +552,21 @@ end
 --
 function M.repos(opts)
   opts = opts or {}
-  local cfg = octo_config.values
   if not opts.login then
     if vim.g.octo_viewer then
       opts.login = vim.g.octo_viewer
     else
-      local remote_hostname = require("octo.utils").get_remote_host()
-      opts.login = require("octo.gh").get_user_name(remote_hostname)
+      local remote_hostname = utils.get_remote_host()
+      local func = backend.get_funcs()["get_user_name"]
+      opts.login = func(remote_hostname)
     end
   end
-  local query = graphql("repos_query", opts.login)
-  utils.info "Fetching repositories (this may take a while) ..."
-  gh.run {
-    args = { "api", "graphql", "--paginate", "--jq", ".", "-f", string.format("query=%s", query) },
-    cb = function(output, stderr)
-      if stderr and not utils.is_blank(stderr) then
-        utils.error(stderr)
-      elseif output then
-        local resp = utils.aggregate_pages(output, "data.repositoryOwner.repositories.nodes")
-        local repos = resp.data.repositoryOwner.repositories.nodes
-        if #repos == 0 then
-          utils.error(string.format("There are no matching repositories for %s.", opts.login))
-          return
-        end
-        local max_nameWithOwner = -1
-        local max_forkCount = -1
-        local max_stargazerCount = -1
-        for _, repo in ipairs(repos) do
-          max_nameWithOwner = math.max(max_nameWithOwner, #repo.nameWithOwner)
-          max_forkCount = math.max(max_forkCount, #tostring(repo.forkCount))
-          max_stargazerCount = math.max(max_stargazerCount, #tostring(repo.stargazerCount))
-        end
-        opts.preview_title = opts.preview_title or ""
-        opts.prompt_title = opts.prompt_title or ""
-        opts.results_title = opts.results_title or ""
-        pickers
-          .new(opts, {
-            finder = finders.new_table {
-              results = repos,
-              entry_maker = entry_maker.gen_from_repo(max_nameWithOwner, max_forkCount, max_stargazerCount),
-            },
-            sorter = conf.generic_sorter(opts),
-            attach_mappings = function(_, map)
-              action_set.select:replace(function(prompt_bufnr, type)
-                open(type)(prompt_bufnr)
-              end)
-              map("i", cfg.picker_config.mappings.open_in_browser.lhs, open_in_browser())
-              map("i", cfg.picker_config.mappings.copy_url.lhs, copy_url())
-              return true
-            end,
-          })
-          :find()
-      end
-    end,
-  }
+  opts.open = open
+  opts.open_in_browser = open_in_browser
+  opts.copy_url = copy_url
+
+  local func = backend.get_funcs()["telescope_repos"]
+  func(opts, octo_config.values)
 end
 
 --
